+++ src/CodeGenerator/src/Generator/GeneratorHelper.php
@@ -63,28 +63,28 @@ class GeneratorHelper
             'SS' => 'Ss',
         ];
         static $ignored = [
-            'GB' => 'Gb',
-            'GiB' => 'Gib',
-            'BFrame' => 'Bframe',
-            'BReference' => 'Breference',
-            'BBox' => 'Bbox',
-            'VCpu' => 'Vcpu',
-            'IInterval' => 'Iinterval',
-            'IFrame' => 'Iframe',
-            'XCoordinate' => 'Xcoordinate',
-            'YCoordinate' => 'Ycoordinate',
-            'XOffset' => 'Xoffset',
-            'YOffset' => 'Yoffset',
-            'XPosition' => 'Xposition',
-            'YPosition' => 'Yposition',
+            'GB',
+            'GiB',
+            'BFrame',
+            'BReference',
+            'BBox',
+            'VCpu',
+            'IInterval',
+            'IFrame',
+            'XCoordinate',
+            'YCoordinate',
+            'XOffset',
+            'YOffset',
+            'XPosition',
+            'YPosition',
         ];
 
         $originalPropertyName = $propertyName;
         $propertyName = strtr($propertyName, $replacements);
 
         if (preg_match('/[A-Z]{2,}/', $propertyName)) {
-            $propertyName = strtr($propertyName, $ignored);
-            if (preg_match('/[A-Z]{2,}/', $propertyName)) {
+            $propertyNameWithoutIgnored = strtr($propertyName, array_fill_keys($ignored, 'Xx'));
+            if (preg_match('/[A-Z]{2,}/', $propertyNameWithoutIgnored)) {
                 throw new \RuntimeException(\sprintf('No camel case property "%s" is not yet implemented', $originalPropertyName));
             }
         }