+++ src/Service/Lambda/CHANGELOG.md
@@ -23,6 +23,7 @@
 
 - Apply new CodingStandard from latest php-cs-fixer.
 - Use a more stable sorting for the list of generated region metadata
+- change case of `executionEnvironmentMemoryGibPerVcpu` method to `executionEnvironmentMemoryGiBPerVCpu`
 
 ## 2.12.0